requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
bcrypt>=4.0.0
PyJWT>=2.8.0
